Demonstrate that we can run tests using `webdriverio`, but it's harder to use.
We can't just rely on setting implicit timeouts like we can with `wd`, so we have to `client.pause(5000)` or otherwise `client.waitFor(...)`, which is pretty verbose.

Also, `appium-test.js` -> `appium-test-wdio.js` for clarity.
